requirements.txt
flask
Pillow
numpy
torchvision